+++ NEWS.md
@@ -1,3 +1,10 @@
+# fastVoteR 0.0.3
+
+* Remove the `data.table` dependency
+* Improve `rank_candidates()` documentation
+* Export internal R functions that wrap C++ code for easier reuse
+* Add extra `check` parameter to all voting functions and `rank_candidates()`
+
 # fastVoteR 0.0.2
 
 * Add contributing guidelines